backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
pymysql==1.1.0
requests==2.31.0
loguru==0.7.2
